+++ tests/menus/_baseMenu/hover.test.js
@@ -129,7 +129,7 @@ describe("BaseMenu", () => {
         initializeMenu(menu);
 
         // Spy on the menu's clearTimeout method.
-        const spy = vi.spyOn(menu, "clearTimeout");
+        const spy = vi.spyOn(menu, "_clearTimeout");
 
         // Simulate the pointerenter event.
         simulatePointerEvent(
@@ -209,7 +209,7 @@ describe("BaseMenu", () => {
           initializeMenu(menu);
 
           // Spy on clearTimeout.
-          const spy = vi.spyOn(menu, "clearTimeout");
+          const spy = vi.spyOn(menu, "_clearTimeout");
 
           // Simulate the pointerleave event.
           simulatePointerEvent(
@@ -233,7 +233,7 @@ describe("BaseMenu", () => {
           initializeMenu(menu);
 
           // Spy on clearTimeout.
-          const spy = vi.spyOn(menu, "clearTimeout");
+          const spy = vi.spyOn(menu, "_clearTimeout");
 
           // Simulate the pointerleave event.
           simulatePointerEvent(
@@ -691,7 +691,7 @@ describe("BaseMenu", () => {
           // Spy on the menu item's clearTimeout method.
           const spy = vi.spyOn(
             menu.elements.submenuToggles[0].elements.controlledMenu,
-            "clearTimeout"
+            "_clearTimeout"
           );
 
           // Simulate the pointerenter event.
@@ -865,7 +865,7 @@ describe("BaseMenu", () => {
           menu.elements.submenuToggles[0].open();
 
           // Spy on the menu's clearTimeout method.
-          const spy = vi.spyOn(menu, "clearTimeout");
+          const spy = vi.spyOn(menu, "_clearTimeout");
 
           // Simulate the pointerenter event.
           simulatePointerEvent(
@@ -961,7 +961,7 @@ describe("BaseMenu", () => {
             // Spy on the menu's clearTimeout method.
             const spy = vi.spyOn(
               menu.elements.submenuToggles[0].elements.controlledMenu,
-              "clearTimeout"
+              "_clearTimeout"
             );
 
             // Simulate the pointerleave event.
@@ -998,7 +998,7 @@ describe("BaseMenu", () => {
             // Spy on the menu's clearTimeout method.
             const spy = vi.spyOn(
               menu.elements.submenuToggles[0].elements.controlledMenu,
-              "clearTimeout"
+              "_clearTimeout"
             );
 
             // Simulate the pointerleave event.
